+++ core/src/main/java/com/fairyband/soak/core/extension/Context.kt
@@ -1,9 +1,13 @@
 package com.fairyband.soak.core.extension
 
+import android.app.Activity
 import android.content.Context
+import android.content.ContextWrapper
+import android.content.Intent
 import android.os.Build
 import android.os.Vibrator
 import android.os.VibratorManager
+import android.provider.Settings
 
 val Context.vibrator: Vibrator
     get() = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
@@ -11,3 +15,16 @@ val Context.vibrator: Vibrator
     } else {
         getSystemService(Context.VIBRATOR_SERVICE) as Vibrator
     }
+
+fun Context.findActivity(): Activity? = when (this) {
+    is Activity -> this
+    is ContextWrapper -> baseContext.findActivity()
+    else -> null
+}
+
+fun Context.openAppNotificationSettings() {
+    val intent = Intent(Settings.ACTION_APP_NOTIFICATION_SETTINGS)
+        .putExtra(Settings.EXTRA_APP_PACKAGE, packageName)
+
+    startActivity(intent)
+}
