--- presentation/src/main/java/com/fairyband/soak/presentation/extensions/Context.kt
@@ -1,20 +1,0 @@
-package com.fairyband.soak.presentation.extensions
-
-import android.app.Activity
-import android.content.Context
-import android.content.ContextWrapper
-import android.content.Intent
-import android.provider.Settings
-
-fun Context.findActivity(): Activity? = when (this) {
-    is Activity -> this
-    is ContextWrapper -> baseContext.findActivity()
-    else -> null
-}
-
-fun Context.openAppNotificationSettings() {
-    val intent = Intent(Settings.ACTION_APP_NOTIFICATION_SETTINGS)
-        .putExtra(Settings.EXTRA_APP_PACKAGE, packageName)
-
-    startActivity(intent)
-}
